pyproject.toml
[tool.poetry]
name = "ao3_tags"
version = "0.1.0"
description = ""
authors = ["Sebastian Heineking <sebastian.heineking@uni-leipzig.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
click = "8.1.7"
elasticsearch = "^8.10.0"
nltk = "^3.8.1"
openai = "^1.30.5"
pandas = "^2.2.3"
pyarrow = "^18.0.0"
pyspark = "^3.5.0"
pyyaml = "^6.0.1"
resiliparse = "*"
scikit-learn= "^1.3.2"
seaborn = "^0.13.2"
spacy = "3.7.2"
statsmodels = "^0.14.1"
textalloc = "^1.0.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
